According to the StAX specs, the requirements for the getPrefix and getNamespacePrefix methods in XMLStreamReader are:
getPrefix: "Returns the prefix of the current event or null if the event does not have a prefix"
getNamespacePrefix: "Returns the prefix for the namespace declared at the index. Returns null if this is the default namespace declaration"

Fixed the implementations of these methods in OMStAXWrapper to return the correct value in the default namespace case (taking into account that in this case Axiom uses the empty string for the prefix).


git-svn-id: https://svn.apache.org/repos/asf/webservices/commons/trunk/modules/axiom@755788 13f79535-47bb-0310-9956-ffa450edef68