Removed the setBuilder method from OMElement. It is an internal method that should not be part of the public API.
Note that strictly speaking, removing that method breaks compatibility with earlier 1.2.x versions. However, using that method in application code would almost inevitably lead to corruption of the internal state of the OMElement. Therefore we can assume that this method is never used in application code and that it is safe to remove it.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axiom/trunk@1503392 13f79535-47bb-0310-9956-ffa450edef68